refactor: split client logic from pages
